added time generation. All events are done save for the cahsier-que system.

+++ src/events/Event.java
@@ -1,9 +1,9 @@
 package events;
 
 public abstract class Event {
-	protected int time;
+	protected double time;
 
-	public int getTime() {
+	public double getTime() {
 		return time;
 	}
